refac: 인터페이스 구현체로 변경
